add method and logic for money deposit

+++ app/src/main/java/javaatm/Costumer.java
@@ -1,7 +1,7 @@
 package javaatm;
 
 public class Costumer {
-  private long balance;
+  private double balance;
   private String name;
 
   public Costumer(String name) {
@@ -9,20 +9,20 @@ public class Costumer {
     this.name = name;
   }
 
-  private void setBalance(long amount) {
+  private void setBalance(double amount) {
     balance = this.balance + amount;
   }
 
-  public long getBalance() {
+  public double getBalance() {
     return this.balance;
   }
 
-  public boolean depositMoney(long amount) {
+  public boolean depositMoney(double amount) {
     setBalance(amount);
     return true;
   }
 
-  public void whitdrawMoney(long amount) {
+  public void whitdrawMoney(double amount) {
     balance -= this.balance - amount;
   }
